src: merge SSLWrap with TLSWrap
`SSLWrap` was introduced in b9a0eb0688104f3619e75e1b9a698080e10d1ec7,
but the only other C++ class to "inherit" from `SSLWrap` went away in
9301b8a9c69d112b98c7d60e074c845d80342b4e. Since all of this API is
already part of `TLSWrap`, move everything to that class. This also
helps move a decent amount of functionality out of `node_crypto` to
a "better" place within C++ land.
